Restructuring for #665 and #681 Improved calculation speed by a tiny fraction.
git-svn-id: https://svn.code.sf.net/p/numere/source/branches/ticket_665_681@1122 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/strings/stringfunctions.hpp
@@ -2096,7 +2096,7 @@ static string strfnc_basetodec(StringFuncArgs& funcArgs)
     }
 
     stream << funcArgs.sArg2;
-    int ret;
+    long long int ret;
     stream >> ret;
 
     return toString(ret);
